Split functions to separate classes adhering to OOP
