--- data/dawn.txt
@@ -1,1 +1,0 @@
-D | 0 | test | by 2024-08-30
