+++ src/main/java/Ui.java
@@ -1,0 +1,29 @@
+import java.io.File;
+import java.util.Scanner;
+
+public class Ui {
+    private Scanner scanner;
+
+    public Ui() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    public void greet() throws DawnException {
+        String divider = "--".repeat(30);
+
+        System.out.println(divider);
+        System.out.println("Dawn 🌙 speaking, what can I do for you?");
+        interact();
+    }
+
+    public void interact() throws DawnException {
+        while (scanner.hasNextLine()) {
+            String command = scanner.next();
+            new Parser(command, scanner.nextLine().trim());
+        }
+    }
+    public void showLoadingError(DawnException ex) {
+        System.out.println(ex);
+    }
+
+}
